requirements.txt
deepspeed==0.17.0
toml
transformers
diffusers>=0.35.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
omegaconf
iopath
termcolor
hydra-core
easydict
ftfy
pytorch-optimizer
wandb
optimum-quanto
